refactor(adr-061): take cli_scan.py off the 2000-line hard cap
`cli_scan.py` sat at exactly 2000 lines with a matching `no_growth`
baseline, while being the most-edited module in the package (25 commits in
six months, against 9 for the next-highest non-scan file). Every change to
`scan` therefore had to be threaded through a file that could not grow by a
line. Two extractions give the responsibility an owner instead:

- `frontends/cli/scan_inputs.py` — turning raw CLI text into typed values or
  rejecting it (budget strings, `--changed` seeds, depth normalisation, ABI3
  floors, source-method resolution). Decides nothing about how a scan runs.
- `frontends/cli/scan_artifact_set.py` — `--artifact-set` member path
  resolution, the comparison-only flag rejection a set cannot honour, and the
  set-shaped renderers, which have no other caller.

2000 -> 1720 lines, and the `debt.yaml` baseline drops with it. Every moved
name is re-exported from `cli_scan.py` under its original private spelling,
so existing call sites, the tests that name them in prose, and any
`monkeypatch.setattr(cli_scan, ...)` resolve exactly as before.

Two findings from `check_architecture.py` shaped the result, and both are
worth recording because they bound how far this kind of split can go today:

1. A first attempt created root `cli_scan_*` siblings. The gate rejected them
   (`frozen-root-family`: ADR-061 D4 forbids a new flat root sibling), which
   is correct — the destination is the `frontends` package, matching where
   `dump` already lives.
2. Inside `frontends`, `unclassified-import` then rejected anything reaching
   an engine module with no ADR-061 owner. That is most of `scan`'s CLI:
   `scan_engine.ScanOutcome` types every report-emission function, and
   `buildsource/crosscheck.ALL_CHECKS` is a runtime dependency of the
   crosscheck parser. A `TYPE_CHECKING` guard does not help — the gate counts
   those imports too, deliberately. So report emission, the crosscheck parser
   and `_run_artifact_set` (which calls it) all stayed in `cli_scan.py`
   rather than being contorted around the rule.

The one classification made is real rather than a workaround: `stable_abi.py`
is a 209-line leaf whose only first-party import is the already-`model`
`stable_abi_data.py`, so it is classified `model`, which is what let the ABI3
floor parser move with its siblings.

Verified: `check_architecture.py` 0 findings, `ruff` and `mypy abicheck/`
clean, and 1506 scan/ABI3-related tests pass.
